Made `smaller_first` and `merge_level` parameters user configurable
