+++ src/main.ts
@@ -1,5 +1,9 @@
 import { createApp } from "vue"
 import "./style.css"
+//* [Toggle]
 import App from "./App.vue"
+/*/
+import App from "./App.debug.vue"
+//*/
 
 createApp(App).mount("#app")
