+++ src/workers/splitRatio.ts
@@ -11,9 +11,19 @@ onmessage = (e: MessageEvent) => {
     e.data.bottleneck_threshold != undefined
       ? new Decimal(e.data.bottleneck_threshold)
       : undefined
+  const merge_level = e.data.merge_level
+  const smaller_first = e.data.smaller_first
   let graph: ConveyorNode[]
   if (e.data.ratio_perms !== true) {
-    graph = main(into, from, max_split, max_merge, bottleneck_threshold)
+    graph = main(
+      into,
+      from,
+      max_split,
+      max_merge,
+      bottleneck_threshold,
+      smaller_first,
+      merge_level,
+    )
   } else {
     graph = main_find_best(
       into,
@@ -21,6 +31,8 @@ onmessage = (e: MessageEvent) => {
       max_split,
       max_merge,
       bottleneck_threshold,
+      smaller_first,
+      merge_level,
     )
   }
